refactored bindings and python test
