+++ bindings/fddp.cpp
