+++ bindings/sqp.cpp
