--- python/py_mim_solvers/unit_tests/run_unit_test.py
@@ -1,6 +1,0 @@
-import unittest
-loader = unittest.TestLoader()
-start_dir = './'
-suite = loader.discover(start_dir)
-runner = unittest.TextTestRunner()
-runner.run(suite)
