+++ python/qp_solvers/py_osqp.py
