+++ python/qp_solvers/qpsolvers.py
