+++ python/qp_solvers/stagewise_qp.py
