+++ python/sqp_cpp.py
