+++ tests/python/test_taichi_sqp.py
@@ -1,6 +1,6 @@
 import pathlib
 import os
-python_path = pathlib.Path('.').absolute().parent
+python_path = pathlib.Path('.').absolute().parent.parent/'python'
 os.sys.path.insert(1, str(python_path))
 import sys
 import crocoddyl
